Updated to version 1.4.1
Remove chmod and stat and add pgrep to ExtComm.
Remove the error in the warn function message.
Remove the test if "EUID == 0".
Change "will not continue" to "Please install the following" in the ExitMessage.
Remove the ! negation when checking for missing apps.
Add an option to kill apper and packagekit.
Test if there are no repositories defined, exit if there is none.
Change the array Reponame to FileNameEnabledRepos.
Add a check if online and if download.opensuse.org is reponding to you.